Relocate ChatHistory from beaker-vue to beaker-ts
