+++ beaker-vue/src/components/panels/ChatHistoryPanel/index.ts
@@ -1,2 +1,3 @@
-export { default as ChatHistoryPanel, type IChatHistory, type IMessage, type IMessageRecord, type ISummaryRecord, type RecordType, type ChatHistoryProps } from './ChatHistoryPanel.vue';
+export { default as ChatHistoryPanel, type ChatHistoryProps } from './ChatHistoryPanel.vue';
+export type { IChatHistory, IMessage, IMessageRecord, ISummaryRecord, RecordType } from '@jataware/beaker-client';
 export { default as ChatHistoryMessage } from './ChatHistoryMessage.vue';
