+++ beaker-vue/src/composables/useNotebookInterface.ts
@@ -1,6 +1,6 @@
 import { ref, computed, watch, nextTick, inject, reactive, type Ref, type ComputedRef } from 'vue';
 import { JupyterMimeRenderer } from '@jataware/beaker-client';
-import type { IBeakerCell, IMimeRenderer } from '@jataware/beaker-client';
+import type { IBeakerCell, IMimeRenderer, IChatHistory } from '@jataware/beaker-client';
 import type { BeakerNotebookComponentType } from '../components/notebook/BeakerNotebook.vue';
 import type { BeakerSessionComponentType } from '../components/session/BeakerSession.vue';
 import { JavascriptRenderer, JSONRenderer, LatexRenderer, MarkdownRenderer, wrapJupyterRenderer, TableRenderer, type BeakerRenderOutput } from '../renderers';
@@ -8,7 +8,6 @@ import { atStartOfInput, atEndOfInput } from '../util';
 import type { NavOption } from '../components/misc/BeakerHeader.vue';
 import { standardRendererFactories } from '@jupyterlab/rendermime';
 import type { IBeakerTheme } from '../plugins/theme';
-import type { IChatHistory } from '../components/panels/ChatHistoryPanel';
 
 interface UseNotebookInterfaceReturn {
     // refs
